fix invalid dereferences in cdaudio_true-unreach-call_true-valid-memsafety.i.cil.c
This commit adds two new entries to the stack, since the program
dereferences one record up and one record down the stack. It also
prevents from out-of-bound access.
That fixes invalid dereferences that led to misclasified
true-unreach-call label. However, there are still memory leaks, so
true-valid-memsafety is really not in place.

Part of the issue #325

Signed-off-by: Marek Chalupa <mchalupa@mail.muni.cz>