ddv-machzwd: fix memory leaks in files labeled as true-valid-memsafety
In all these files there's one call to malloc in the request_region function
and the allocated memory is then leaked. Since the call is only one,
add a static global variable that remembers the allocation and free
the memory in the release_region function. Reason why the function is
global is because release_region does not take the address as an
argument (it takes some other identifier that we would need to fake
somehow).

This fixes few files from #333

Signed-off-by: Marek Chalupa <mchalupa@mail.muni.cz>